+++ tests/service/ProjectOptionsTests.fs
@@ -1,0 +1,141 @@
+﻿#if INTERACTIVE
+#r "../../bin/v4.5/FSharp.Compiler.Service.dll"
+#r "../../packages/NUnit.2.6.3/lib/nunit.framework.dll"
+#load "FsUnit.fs"
+#load "Common.fs"
+#else
+module FSharp.Compiler.Service.Tests.ProjectOptionsTests
+#endif
+
+let runningOnMono = try System.Type.GetType("Mono.Runtime") <> null with e ->  false
+
+open System
+open NUnit.Framework
+open FsUnit
+open Microsoft.FSharp.Compiler.SourceCodeServices
+
+open FSharp.Compiler.Service.Tests.Common
+
+#if FX_ATLEAST_45
+
+let checkOption (opts:string[]) s = 
+    let found = "Found '"+s+"'"
+    (if opts |> Array.exists (fun o -> o.EndsWith(s)) then found else "Failed to find '"+s+"'")
+       |> shouldEqual found
+
+let checkOptionNotPresent (opts:string[]) s = 
+    let found = "Found '"+s+"'"
+    let notFound = "Did not expect to find '"+s+"'"
+    (if opts |> Array.exists (fun o -> o.EndsWith(s)) then found else notFound)
+       |> shouldEqual notFound
+
+[<Test>]
+let ``Project file parsing example 1 Default Configuration`` () = 
+  // BUG - see https://github.com/fsharp/FSharp.Compiler.Service/issues/237
+//  if not runningOnMono then 
+
+    let projectFile = __SOURCE_DIRECTORY__ + @"/FSharp.Compiler.Service.Tests.fsproj"
+    let options = checker.GetProjectOptionsFromProjectFile(projectFile)
+
+    printfn "vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv" 
+    printfn "PROJ FILE %s" projectFile
+    printfn "%A" options.OtherOptions
+    printfn "^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^" 
+
+    checkOption options.OtherOptions "FSharp.Compiler.Service.dll"
+    checkOption options.OtherOptions "FileSystemTests.fs"
+    checkOption options.OtherOptions "--define:TRACE"
+    checkOption options.OtherOptions "--define:DEBUG"
+    checkOption options.OtherOptions "--flaterrors"
+    checkOption options.OtherOptions "--simpleresolution"
+    checkOption options.OtherOptions "--noframework"
+
+[<Test>]
+let ``Project file parsing example 1 Release Configuration`` () = 
+  // BUG - see https://github.com/fsharp/FSharp.Compiler.Service/issues/237
+//  if not runningOnMono then 
+    let projectFile = __SOURCE_DIRECTORY__ + @"/FSharp.Compiler.Service.Tests.fsproj"
+    // Check with Configuration = Release
+    let options = checker.GetProjectOptionsFromProjectFile(projectFile, [("Configuration", "Release")])
+
+    printfn "vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv" 
+    printfn "PROJ FILE %s" projectFile
+    printfn "%A" options.OtherOptions
+    printfn "^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^" 
+    checkOption options.OtherOptions "FSharp.Compiler.Service.dll"
+    checkOption options.OtherOptions "FileSystemTests.fs"
+    checkOption options.OtherOptions "--define:TRACE"
+    checkOptionNotPresent options.OtherOptions "--define:DEBUG"
+    checkOption options.OtherOptions "--debug:pdbonly"
+
+[<Test>]
+let ``Project file parsing example 1 Default configuration relative path`` () = 
+  // BUG - see https://github.com/fsharp/FSharp.Compiler.Service/issues/237
+//  if not runningOnMono then 
+
+    let projectFile = "FSharp.Compiler.Service.Tests.fsproj"
+    Environment.CurrentDirectory <- __SOURCE_DIRECTORY__
+
+
+    let options = checker.GetProjectOptionsFromProjectFile(projectFile)
+    //let options = FSharpProjectFileInfo.Parse(projectFile)
+    //options.FullPath
+    //options.Options
+
+    printfn "vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv" 
+    printfn "PROJ FILE %s" projectFile
+    printfn "%A" options.OtherOptions
+    printfn "^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^" 
+
+    checkOption options.OtherOptions "FSharp.Compiler.Service.dll"
+    checkOption options.OtherOptions "FileSystemTests.fs"
+    checkOption options.OtherOptions "--define:TRACE"
+    checkOption options.OtherOptions "--define:DEBUG"
+    checkOption options.OtherOptions "--flaterrors"
+    checkOption options.OtherOptions "--simpleresolution"
+    checkOption options.OtherOptions "--noframework"
+
+[<Test>]
+let ``Project file parsing VS2013_FSharp_Portable_Library_net45``() = 
+  // BUG - see https://github.com/fsharp/FSharp.Compiler.Service/issues/237
+  if not runningOnMono then 
+    let projectFile = __SOURCE_DIRECTORY__ + @"/../projects/Sample_VS2013_FSharp_Portable_Library_net45/Sample_VS2013_FSharp_Portable_Library_net45.fsproj"
+    let options = checker.GetProjectOptionsFromProjectFile(projectFile, [])
+
+    printfn "vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv" 
+    printfn "PROJ FILE %s" projectFile
+    printfn "%A" options.OtherOptions
+    printfn "^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^" 
+
+    checkOption options.OtherOptions "--targetprofile:netcore"
+    checkOption options.OtherOptions "--tailcalls-"
+
+    checkOption options.OtherOptions "FSharp.Core.dll"
+    checkOption options.OtherOptions "Microsoft.CSharp.dll"
+    checkOption options.OtherOptions "System.Runtime.dll"
+    checkOption options.OtherOptions "System.Net.Requests.dll"
+    checkOption options.OtherOptions "System.Xml.XmlSerializer.dll"
+
+[<Test>]
+let ``Project file parsing Sample_VS2013_FSharp_Portable_Library_net451_adjusted_to_profile78``() = 
+  // BUG - see https://github.com/fsharp/FSharp.Compiler.Service/issues/237
+  if not runningOnMono then 
+    let projectFile = __SOURCE_DIRECTORY__ + @"/../projects/Sample_VS2013_FSharp_Portable_Library_net451_adjusted_to_profile78/Sample_VS2013_FSharp_Portable_Library_net451.fsproj"
+    let options = checker.GetProjectOptionsFromProjectFile(projectFile, [])
+
+    printfn "vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv" 
+    printfn "PROJ FILE %s" projectFile
+    printfn "%A" options.OtherOptions
+    printfn "^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^" 
+    checkOption options.OtherOptions "--targetprofile:netcore"
+    checkOption options.OtherOptions "--tailcalls-"
+
+    checkOption options.OtherOptions "FSharp.Core.dll"
+    checkOption options.OtherOptions "Microsoft.CSharp.dll"
+    checkOption options.OtherOptions "System.Runtime.dll"
+    checkOption options.OtherOptions "System.Net.Requests.dll"
+    checkOption options.OtherOptions "System.Xml.XmlSerializer.dll"
+
+#endif
+
+